removed the existing sink and source tasks from executor and refactored packages accordinly
