+++ twister2/executor/src/java/edu/iu/dsc/tws/executor/api/ParallelOperationFactory.java
@@ -29,18 +29,18 @@ import java.util.logging.Logger;
 import edu.iu.dsc.tws.common.config.Config;
 import edu.iu.dsc.tws.comms.core.TaskPlan;
 import edu.iu.dsc.tws.comms.op.Communicator;
-import edu.iu.dsc.tws.executor.comm.operations.batch.AllReduceBatchOperation;
-import edu.iu.dsc.tws.executor.comm.operations.batch.BroadcastBatchOperation;
-import edu.iu.dsc.tws.executor.comm.operations.batch.GatherBatchOperation;
-import edu.iu.dsc.tws.executor.comm.operations.batch.KeyedReduceBatchOperation;
-import edu.iu.dsc.tws.executor.comm.operations.batch.PartitionBatchOperation;
-import edu.iu.dsc.tws.executor.comm.operations.batch.ReduceBatchOperation;
-import edu.iu.dsc.tws.executor.comm.operations.streaming.AllReduceStreamingOperation;
-import edu.iu.dsc.tws.executor.comm.operations.streaming.BroadcastStreamingOperation;
-import edu.iu.dsc.tws.executor.comm.operations.streaming.GatherStreamingOperation;
-import edu.iu.dsc.tws.executor.comm.operations.streaming.KeyedReduceStreamingOperation;
-import edu.iu.dsc.tws.executor.comm.operations.streaming.PartitionStreamingOperation;
-import edu.iu.dsc.tws.executor.comm.operations.streaming.ReduceStreamingOperation;
+import edu.iu.dsc.tws.executor.comms.batch.AllReduceBatchOperation;
+import edu.iu.dsc.tws.executor.comms.batch.BroadcastBatchOperation;
+import edu.iu.dsc.tws.executor.comms.batch.GatherBatchOperation;
+import edu.iu.dsc.tws.executor.comms.batch.KeyedReduceBatchOperation;
+import edu.iu.dsc.tws.executor.comms.batch.PartitionBatchOperation;
+import edu.iu.dsc.tws.executor.comms.batch.ReduceBatchOperation;
+import edu.iu.dsc.tws.executor.comms.streaming.AllReduceStreamingOperation;
+import edu.iu.dsc.tws.executor.comms.streaming.BroadcastStreamingOperation;
+import edu.iu.dsc.tws.executor.comms.streaming.GatherStreamingOperation;
+import edu.iu.dsc.tws.executor.comms.streaming.KeyedReduceStreamingOperation;
+import edu.iu.dsc.tws.executor.comms.streaming.PartitionStreamingOperation;
+import edu.iu.dsc.tws.executor.comms.streaming.ReduceStreamingOperation;
 import edu.iu.dsc.tws.executor.core.CommunicationOperationType;
 import edu.iu.dsc.tws.task.api.Operations;
 import edu.iu.dsc.tws.task.graph.Edge;
