+++ twister2/executor/src/java/edu/iu/dsc/tws/executor/comms/batch/KeyedReduceBatchOperation.java
@@ -9,7 +9,19 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-package edu.iu.dsc.tws.executor.comm.operations.streaming;
+
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+package edu.iu.dsc.tws.executor.comms.batch;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -30,12 +42,12 @@ import edu.iu.dsc.tws.executor.api.EdgeGenerator;
 import edu.iu.dsc.tws.executor.util.Utils;
 import edu.iu.dsc.tws.task.api.IMessage;
 
-public class KeyedReduceStreamingOperation extends AbstractParallelOperation {
-  private static final Logger LOG = Logger.getLogger(KeyedReduceStreamingOperation.class.getName());
+public class KeyedReduceBatchOperation extends AbstractParallelOperation {
+  private static final Logger LOG = Logger.getLogger(KeyedReduceBatchOperation.class.getName());
 
   protected DataFlowMultiReduce op;
 
-  public KeyedReduceStreamingOperation(Config config, Communicator network, TaskPlan tPlan) {
+  public KeyedReduceBatchOperation(Config config, Communicator network, TaskPlan tPlan) {
     super(config, network, tPlan);
   }
 
@@ -56,7 +68,11 @@ public class KeyedReduceStreamingOperation extends AbstractParallelOperation {
 
   @Override
   public boolean progress() {
-    return op.progress();
+    return op.progress() && hasPending();
+  }
+
+  public boolean hasPending() {
+    return !op.isComplete();
   }
 
   /**
