+++ twister2/executor/src/java/edu/iu/dsc/tws/executor/comms/streaming/BroadcastStreamingOperation.java
@@ -9,7 +9,19 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-package edu.iu.dsc.tws.executor.comm.operations.batch;
+
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+package edu.iu.dsc.tws.executor.comms.streaming;
 
 import java.util.List;
 import java.util.Map;
@@ -29,11 +41,11 @@ import edu.iu.dsc.tws.executor.util.Utils;
 import edu.iu.dsc.tws.task.api.IMessage;
 import edu.iu.dsc.tws.task.api.TaskMessage;
 
-public class BroadcastBatchOperation extends AbstractParallelOperation {
-  private static final Logger LOG = Logger.getLogger(BroadcastBatchOperation.class.getName());
+public class BroadcastStreamingOperation extends AbstractParallelOperation {
+  private static final Logger LOG = Logger.getLogger(BroadcastStreamingOperation.class.getName());
   private DataFlowBroadcast op;
 
-  public BroadcastBatchOperation(Config config, Communicator network, TaskPlan tPlan) {
+  public BroadcastStreamingOperation(Config config, Communicator network, TaskPlan tPlan) {
     super(config, network, tPlan);
   }
 
@@ -54,11 +66,7 @@ public class BroadcastBatchOperation extends AbstractParallelOperation {
 
   @Override
   public boolean progress() {
-    return op.progress() && hasPending();
-  }
-
-  public boolean hasPending() {
-    return !op.isComplete();
+    return op.progress();
   }
 
   public class BcastReceiver implements MessageReceiver {
